# backend/docs/openapi.yaml
components:
...
  securitySchemes:
    apiKeyAuth:
      description: API key for device authentication. Provide the API key as a Bearer
        token.
      in: header
      name: Authorization
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    name: MOTO Support
  description: API for the MOTO school management system
  title: MOTO API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/active/groups:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/active/groups
      tags:
      - Active
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/active/groups
      tags:
      - Active
...
  /api/active/visits:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/active/visits
      tags:
      - Active
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/active/visits
      tags:
      - Active
...
  /api/groups:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/groups
      tags:
      - Groups
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/groups
      tags:
      - Groups
  /api/groups/{id}:
    delete:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: DELETE /api/groups/{id}
      tags:
      - Groups
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/groups/{id}
      tags:
      - Groups
    put:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: PUT /api/groups/{id}
      tags:
      - Groups
...
  /api/staff:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/staff
      tags:
      - Staff
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/staff
      tags:
      - Staff
  /api/staff/{id}:
    delete:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: DELETE /api/staff/{id}
      tags:
      - Staff
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/staff/{id}
      tags:
      - Staff
    put:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: PUT /api/staff/{id}
      tags:
      - Staff
...
  /api/students:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/students
      tags:
      - Students
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/students
      tags:
      - Students
  /api/students/{id}:
    delete:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: DELETE /api/students/{id}
      tags:
      - Students
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/students/{id}
      tags:
      - Students
    put:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: PUT /api/students/{id}
      tags:
      - Students
...
  /api/users:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/users
      tags:
      - Users
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/users
      tags:
      - Users
  /api/users/{id}:
    delete:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: DELETE /api/users/{id}
      tags:
      - Users
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/users/{id}
      tags:
      - Users
    put:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: PUT /api/users/{id}
      tags:
      - Users
...
  /api/users/{id}/profile:
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/users/{id}/profile
      tags:
      - Users
...
servers:
- description: API Base URL
  url: /api